Renamed types module to avoid conflict with stdlib
